+++ plugins/modules/lsnsipalgcall.py
@@ -40,7 +40,7 @@ options:
     description:
       - Call ID for the SIP call.
   nodeid:
-    type: float
+    type: int
     description:
       - Unique number that identifies the cluster node.
 extends_documentation_fragment: netscaler.adc.netscaler_adc
